Make component to share logic between smart search dialog and column editor. To be refactored soon bc not pretty.
